+++ examples/with-thirdpartypasswordless-electron/src/Footer/index.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+export default function Footer(): JSX.Element {
+    return (
+        <div
+            style={{
+                display: "flex",
+                width: "100%",
+                height: "80px",
+                backgroundColor: "#000000",
+                alignItems: "center",
+                justifyContent: "center",
+                alignSelf: "flex-end",
+                color: "#ffffff",
+                fontWeight: "bold",
+            }}>
+            React Demo app. Made with ❤️ using supertokens.com
+        </div>
+    );
+}