+++ lib/ts/recipe/session/recipe.ts
@@ -22,7 +22,7 @@ import { isTest } from "../../utils";
 import { InputType, RecipeEvent, RecipeEventWithSessionContext, SessionContextType } from "./types";
 import sessionSdk from "supertokens-website";
 
-type ConfigType = InputType & { recipeId: string; appInfo: NormalisedAppInfo };
+type ConfigType = InputType & { recipeId: string; appInfo: NormalisedAppInfo; enableDebugLogs: boolean };
 
 export default class Session extends RecipeModule<unknown, unknown, unknown, any> {
     static instance?: Session;
@@ -114,7 +114,7 @@ export default class Session extends RecipeModule<unknown, unknown, unknown, any
     };
 
     private async getSessionContext({ action }: RecipeEvent): Promise<SessionContextType> {
-        if (action === "SESSION_CREATED" || action === "REFRESH_SESSION") {
+        if (action === "SESSION_CREATED" || action === "REFRESH_SESSION" || action === "ACCESS_TOKEN_PAYLOAD_UPDATED") {
             const [userId, accessTokenPayload] = await Promise.all([
                 this.getUserId(),
                 this.getAccessTokenPayloadSecurely(),
@@ -144,11 +144,12 @@ export default class Session extends RecipeModule<unknown, unknown, unknown, any
     }
 
     static init(config?: InputType): CreateRecipeFunction<unknown, unknown, unknown, any> {
-        return (appInfo: NormalisedAppInfo): RecipeModule<unknown, unknown, unknown, any> => {
+        return (appInfo: NormalisedAppInfo, enableDebugLogs: boolean): RecipeModule<unknown, unknown, unknown, any> => {
             Session.instance = new Session({
                 ...config,
                 appInfo,
                 recipeId: Session.RECIPE_ID,
+                enableDebugLogs,
             });
             return Session.instance;
         };