+++ lib/ts/recipe/session/types.ts
@@ -18,7 +18,7 @@ import OverrideableBuilder from "supertokens-js-override";
 
 export type RecipeEvent =
     | {
-          action: "SIGN_OUT" | "REFRESH_SESSION" | "SESSION_CREATED";
+          action: "SIGN_OUT" | "REFRESH_SESSION" | "SESSION_CREATED" | "ACCESS_TOKEN_PAYLOAD_UPDATED";
       }
     | {
           action: "UNAUTHORISED";