+++ lib/ts/types.ts
@@ -19,6 +19,8 @@ import { CSSObject } from "@emotion/react/types/index";
 import { ComponentClass, PropsWithChildren } from "react";
 import { NormalisedConfig as NormalisedRecipeModuleConfig } from "./recipe/recipeModule/types";
 import { TranslationFunc, TranslationStore } from "./translation/translationHelpers";
+import { CookieHandlerInput } from "supertokens-website/utils/cookieHandler/types";
+import { WindowHandlerInput } from "supertokens-website/utils/windowHandler/types";
 
 /*
  * Recipe Module Manager Config Types.
@@ -34,6 +36,10 @@ export type SuperTokensConfig = {
      * List of recipes for authentication and session management.
      */
     recipeList: CreateRecipeFunction<any, any, any, any>[];
+
+    cookieHandler?: CookieHandlerInput;
+
+    windowHandler?: WindowHandlerInput;
 
     languageTranslations?: {
         /*
@@ -64,10 +70,12 @@ export type SuperTokensConfig = {
          */
         translationFunc?: TranslationFunc;
     };
+    enableDebugLogs?: boolean;
 };
 
 export type CreateRecipeFunction<T, S, R, N extends NormalisedRecipeModuleConfig<T, S, R>> = (
-    appInfo: NormalisedAppInfo
+    appInfo: NormalisedAppInfo,
+    enableDebugLogs: boolean
 ) => RecipeModule<T, S, R, N>;
 
 export type AppInfoUserInput = {