+++ CHANGELOG.md
@@ -5,11 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [unreleased]
-
--   Adds an example app with emailpassword + vercel
-
-## [0.20.5] - 2022-05-04
+## [0.21.2] - 2022-05-13
 
 ### Fixes
 
@@ -18,6 +14,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 -   Updated react dependency to react 18
+
+## [0.21.1] - 2022-05-13
+
+-   Adds debug logging
+
+## [0.21.0] - 2022-05-11
+
+### Changes
+
+-   Now updating the session context when the access token payload is updated
+-   Updated supertokens-webiste dependency to ^11.0.0
+
+### Added
+
+## [0.20.5] - 2022-05-10
+
+### Adds
+
+-   An example app with emailpassword + vercel
+-   A new config property `cookieHandler` that allows for custom handling when the SDK reads/writes cookies
+-   A new config property `windowHandler` that allows for custom handling when the SDK uses any functions from the Window API.
+-   An example app with thirdpartypasswordless + Electron
 
 ## [0.20.4] - 2022-04-11